Remove new cyclic dependencies
