+++ backend/apps/iamstudent/forms.py
@@ -861,12 +861,3 @@ def get_form_helper_filter():
     return helper
 
 
-from .models_persistent_filter import StudentListFilterModel
-
-
-class StudentListFilterModelForm(forms.ModelForm):
-    class Meta:
-        model = StudentListFilterModel
-        labels = form_labels
-        labels["braucht_bezahlung"] = _("Vergütung möglich")
-        exclude = []
